requirements.txt
streamlit>=1.31.0
langchain-ollama>=0.0.3
langchain-core>=0.1.0
python-dateutil>=2.8.2
python-logging>=0.4.9.6
dataclasses>=0.6
typing>=3.7.4.3